end of concrete_solve
After seeing how these tutorials turned out, I took it upon myself to figure out how to make it so that DiffEq code could use adjoints with literally no change and is safe. Turns out it works. This PR removes `concrete_solve` and makes it so that it's all `solve` usage, since `concrete_solve` is now deprecated in DiffEqBase v6.36. The manifest might need an update, and my computer is the absolute worst case right now for that so I just directly edited the rendered results.