[buildkite] Update S3 credential handling
We were alerted that our S3 credentials had leaked in a build that dumps
the environment when it errors out [0]. Because `TuringTutorials` uses
the same infrastructure as `SciMLBenchmarks`, we have to update things
here as well.

 This PR rotates to a new key, and also wipes out the key in the
environment during the build itself, so that sensitive values won't get
written out in debugging information like that.

[0] https://github.com/SciML/SciMLBenchmarksOutput/blob/8d80b1840228c63e8c6fdc597049211a8b916dea/markdown/DynamicalODE/single_pendulums.md